requirements/development.txt
-r common.txt

# Documentation
# -------------------------------------
isort==5.6.4
black==20.8b1

# Debugging
# -------------------------------------
django-debug-toolbar==3.2
ipdb==0.13.4

# Testing and coverage
# -------------------------------------
pytest==6.2.0
pytest-django==4.1.0
pytest-cov==2.10.1
django-dynamic-fixture==3.1.1
flake8-mypy==17.8.0
pytest-mock==3.3.1

# Code quality
#--------------------------------------
pre-commit==2.9.3
